refactor: split query tests into different files
